+++ functions/lib/index.js
@@ -1,12 +1,11 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.addTile = exports.createColony = exports.getColony = exports.makeUppercase = void 0;
+exports.addTile = exports.createColony = exports.makeUppercase = void 0;
 const firestore_1 = require("firebase-functions/v2/firestore");
 const logger = require("firebase-functions/logger");
 const admin = require("firebase-admin");
 // Import functions from other files
 const colony_1 = require("./colony");
-Object.defineProperty(exports, "getColony", { enumerable: true, get: function () { return colony_1.getColony; } });
 Object.defineProperty(exports, "createColony", { enumerable: true, get: function () { return colony_1.createColony; } });
 const tiles_1 = require("./tiles");
 Object.defineProperty(exports, "addTile", { enumerable: true, get: function () { return tiles_1.addTile; } });